Doc cleanup prior to PyCon 2019 (#38)
* General Cleanup of Docs Prion to PyCon 2019

Went through a general cleanup of the content. Also reduced the README to point to the ReadTheDocs content (which should be updated when this PR is merged).

* Fixing line 258 length

This is to fix a flake8 issue.

* Fix for flake8  error

There were 4 lines that either had extra whitespace, or that were too long.

* Adding # noqa to comment lines

I'm adding # noqa to comment lines this to preserve the whitespaces

* Adding # noqa to preserve doc comments.

Adding # noqa to preserve doc comments to prevent unnecessary flake8 errors.